+++ run.py
@@ -10,6 +10,6 @@ if sys.stdout is None:
 if sys.stderr is None:
     sys.stderr = open(os.devnull, 'w')
 
-print('Starting WiprFlow...')
+print('Starting WhisperEdge...')
 load_dotenv()
 subprocess.run([sys.executable, os.path.join('src', 'main.py')])